requirements.txt
tqdm>=4.65.0
requests>=2.31.0
py7zr>=0.20.5
tensorflow>=2.13.0
opencv-python>=4.8.0
scikit-learn>=1.3.0
matplotlib>=3.7.2
pillow>=10.0.0
seaborn>=0.12.2
opencv-python-headless>=4.8.0
numpy>=1.24.3
torch>=2.0.1
torchvision>=0.15.2
pandas>=2.0.3
scipy>=1.11.2
albumentations>=1.3.1
grad-cam>=1.4.8
timm>=0.9.2
scikit-image>=0.21.0
wandb>=0.15.8  # for experiment tracking